backend/requirements.txt
boto3
celery
celery[redis]
click
flask
flask-cors
flask-jwt-extended
flask-migrate
flask-smorest
flask-sqlalchemy
google-auth
google-generativeai
gunicorn
lxml[html_clean]
marshmallow
newspaper4k
passlib
psycopg2-binary
python-dotenv
redis
requests
resend
slack_sdk
uuid6
